Migrate test_range to Travis suite
