+++ tests/selenium/test_range.py
@@ -1,11 +1,8 @@
-import pytest
-
-from pages.treeherder import TreeherderPage
+from pages.treeherder import Treeherder
 
 
-@pytest.mark.nondestructive
-def test_set_as_top_of_range(base_url, selenium):
-    page = TreeherderPage(selenium, base_url).open()
+def test_set_as_top_of_range(base_url, selenium, test_job):
+    page = Treeherder(selenium, base_url).open()
     result_sets = page.result_sets
     datestamp = result_sets[1].datestamp
     assert result_sets[0].datestamp != datestamp
@@ -13,9 +10,8 @@ def test_set_as_top_of_range(base_url, selenium):
     assert page.result_sets[0].datestamp == datestamp
 
 
-@pytest.mark.nondestructive
-def test_set_as_bottom_of_range(base_url, selenium):
-    page = TreeherderPage(selenium, base_url).open()
+def test_set_as_bottom_of_range(base_url, selenium, test_job):
+    page = Treeherder(selenium, base_url).open()
     result_sets = page.result_sets
     datestamp = result_sets[-2].datestamp
     assert result_sets[-1].datestamp != datestamp
